Add writeTime as an optional parameter when writing stream
This commit adds a character string 'writeTime' as an optional paramter
to the mpas_stream_mgr_write routine, which can define the time you want
the data to be written at. This is useful to help manipulate the
filename of the data, if you want it at a different time stamp than the
current time of the clock.

If the writeTime is not passed in, it defaults to the current time of
the clock associated with the stream manager.

A test is also added to the test core to help exercise this
functionality.