Make compset match more specific for ROF_NCPL
Mariana agreed with me that the match on "_DATM.*_POP2" is really just
meant to apply for DROF, so I have added that explicitly.

Probably most of the other matches in this section on ROF_NCPL are only
supposed to apply to DROF as well, but we weren't as certain of that. We
have left those as is for now, because they don't affect the J compset
(which is what is motivating this change).

Test suite: None
Test baseline: N/A
Test namelist changes: N/A
Test status: N/A

Fixes: None

User interface changes?: No

Code review: None