Separate RK4 into subroutines to allow external/internal loops
This commit splits RK4 into a few subroutines which work on blocks,
rather than having the code within a block loop in the main routine.
This allows block loops to be split into external and internal loops
without duplicating much of the code.

Additionally, this removes some redundant computations and simplifies
some of the loops.