Handle times better in nightly testing infrastructure.
On some machines, the time gap between when tests started and when
they finished was so long that CTest was getting confused during
the submit phase and ignoring the results we wanted to submit.

This commit changes the timestamp of the results from the time tests
start to the time when they finish. It also leverages a new environment
variable in jenkins_script to provide a much more accurate estimate of
the total time spent in the test.

[BFB]