implementing workaround for PGI array indexing algorithm limitation
The module private arrays aer_cb and aer in the hetfrz_classnum_cam are
dimensioned as

aer_cb(pcols,pver,ncnst,begchunk:endchunk)
aer(pcols,pver,ncnst,begchunk:endchunk)

For one of the target compsets, pcols = 16, pver = 72, and ncnst = 14.
When begchunk is greater than approximately 133,000, the index calculation
when using the PGI compiler overflows 32 bit (signed) integers, and
generates a runtime segmentation fault. This is independent of the actual
range (begchunk:endchunk), which is typically quite small.

By changing the dimensions to

aer_cb(pcols,pver,ncnst,0:(endchunk-begchunk))
aer(pcols,pver,ncnst,0:(endchunk-begchunk))

and using

lchnk_zb = state%lchnk - begchunk

when indexing, this error can be avoided. PGI suggests instead building
with the -mcmodel=medium flag, but this causes CAM to fail with a different
error when applied globally, and hangs in this module when applied only
to the module.

Note that other arrays dimensioned with (begchunk:endchunk) are safe for
much larger begchunk values.