Widen MISMIP domains that were too narrow
The 500m (and 250m?) resolution MISMIP3D domains were too narrow such that
the ice goes up to the edge, causing trouble for the solver.  This
commit widens those domains so there is a healthy buffer along the east
and west.

This commit also increases the initial ice thickness so that it is grounded
since the specified initial thickness of 10m results in a big iceberg
that causes velocity solver issues.