Fixes issue in nonlocal transport of CVMix
Presently the forcing of the first model layer of temperature and
salinity is in two parts, the direct forcing, and a dilution effect.

Currently, only the direct forcing of temperature and salinity is
included in the non-local transport portion of CVMix.  This commit
introduces a new variable nonLocalSurfaceTracerFlux that combines
the two forcings and sends this array to the nonlocal transport
instead of activeTracersSurfaceFlux