Add subroutine for deepening critical passages
This merge adds a package for fields related to critical passages
and the fields for masks, depths and level number of each critical
passage.  A subroutine is added for deepening the bathymetry where
it is shallower than the minimum depth read in for each critical
passage.  Ocean cells are no longer marked as land if the observed
bottom depth is above sea level.