Fix a bug to enable BC/dust deposition on snow/ice with marine organics
The new marine organic aerosol chemistry option is enabled with a new
preprocessor directive, and the new treatment of BC deposition to
ice/snow uses chemistry preprocessor directives to determine which
scheme to use.  The preprocessor directive enabling the marine organic
aerosol option, MODAL_AERO_4MODE_MOM, was missing from the code, such
that BC/dust deposition to ice would not be treated in simulations
with marine organic aerosol enabled.  This commit adds the
pre-processor directive so that they will both be active at the same
time.

[CC] for configurations that enable both marine organics and BC/dust
deposition in snow/ice, including the FC5AV1C, WCYCL, and CRYO
compsets.