Enable BC/dust deposition to snow/ice with 9-mode MAM
Enable new treatment of BC/dust deposition to snow/ice to be turned on
in combination with the more detailed, 9-mode MAM treatment of marine
organic aerosol.

The 9-mode aerosol treatment is not used in any currently tested
compsets, so this commit should be BFB with all current tests.  It
should be climate-changing when using MAM9 configurations of the
model.

[BFB]
[FCC]