Comment out call to check_topo
It's important for topographic heights to be within bounds in order for
the limiter to be applied correctly. However, this currently isn't the
case for some of the old TG forcing data. At a glance, it looks like the
problems are just outside of Greenland, so this should be okay. When we
have new TG forcing data, we should try uncommenting this call to
check_topo.

Test suite: cime unit tests
Test baseline: N/A
Test namelist changes: N/A
Test status: pass

Fixes: None

User interface changes?: No

Code review: None