Regional statistics analysis member.
This analysis member performs regional statistics over masked areas,
such that it requires an external mask stream that defines the masks.
The mask stream input requires a (regions, elements) 2D array of integers,
with (0 or 1s) indicating if mesh elements are in or out of a region.

The reduction operations that can be applied are min, max, and average,
supporting cell and vertex reductions and optionally also applying
a vertical mask as well, such as cellMask in the ocean core, if the
input has a vertical dimension, such as nVertLevels in the ocean core.

This AM package also includes a RPN (reverse polish notation, otherwise
known as a stack) calculator for creating simple fields and values. In
particular, it was added to easily calculate cell volume from area
and depth in the ocean core without having to introduce a special case
in the regional statistics AM.

It is a 0D, 1D, and 2D vector calculator with 4 functions, which allows
for dimension broadcasting operations. Future extensions can allow for
reductions and other operations.