Add -no-opt-dynamic-align to Cori Intel FFlags
Enforce same code to be executed regardless of data alignment with
-no-opt-dynamic-align. By default, the compiler may generate multiple
code paths to execute depending on the alignment of data in order to
improve performance which may affect the consistency of floating-point
calculations. To disable this behavior, use -no-opt-dynamic-align.

This fixes ERS tests that were previously failing on Cori.