Fix an out of bounds error with rSurfaceLayer
This commit fixes an issue caused when rSurfaceLayer = nVertLevels. In
this case, k is set to be int(rSurfaceLayer)+1, which is an out of
bounds access and causes a seg fault.

This causes the max of k to be maxLevelCell(iCell) so as to prevent out
of bounds accesses, or accessing garbage from cells below active ocean
cells.