Add uReconstructX to input file for PxxS experiment if field exists
At high resolution, Albany was having problems with the first velocity
solve when the initial guess was 0.  This change update the script that
sets up the input file for the perturbation experiment so that if
uReconstructX is present in the Stnd file, then it will be transferred
to the PxxS input file.  Note that to get uReconstructX plus the other
required fields in one file will likely require some manual ncks operations.