Fixes a bug in variable shortwave scheme
This commit fixes two issues in the Ohlman 00 shortwave scheme.  First,
if clearSkyRadiation is ever zero, a segFault will occur, due to division
by zero.  A small positive number is added to prevent this.  Second, there is
presently no check for the necessary data arrays prior to computing extinction
coefficients.  If config_sw_absorption_type is set to ohlman00 and no data arrays
are provided, the code will attempt to compute extinction coefficients.  In this
commit a check is added for these arrays, if any are missing, a warning is
printed and the code is not run.