Modify the algorithm for Haney-number init
This merge significantly modifies the algorithm for Haney-number
constrained initialization of vertical layers.  The gist of the change
is that a stretch field is introduced indicating the deviation of
the vertical grid from z-level.  The algorithm now tries to keep
the stretch field smooth, rather than layer thickness, leading to
a smoother transition to z-level outside the region where the
constraint is applied.

Several constants have been added and removed from the registry
related to the change in algorithm.  These will be updated in the
individual testcases as part of a separate commit.

The algorithm for computing the Haney number at cells and edges
was also modified to include computing the Haney number in the top
layer using the SSH.