update to st_archive to correctly check if interim restart file duplicates should be removed
The st_archive was incorrectly removing duplicate rpointer files in the archive/$case/rest/[datename]
directories to do an incorrect string comparison in the perl code. Also updated the archive.xml
to match on rs1.* suffices for the data models.

Test suite: A compset test with toggling of restart settings in order to debug
Test baseline: N/A
Test namelist changes: N/A
Test status: bit for bit

Fixes: #445

Code review: