Allow pass-through of any field in time series stats AM.
It allows things like "xtime" and "mesh" to show up in the output
stream that the time series stats AM scans for input fields.

It checks if the field is real and has a time dimension, if so
it will apply time series stats to the data. If not, it just passes
the field data to the output stream without modification (i.e., no
time dimension reduction applied), with a warning.