fix calc. of max and min thickness
This commit fixes the logic for calculating the max and min thickness
values when using regions masks (the application of the regions mask
was accidentally left off of the original set of commits).

I've verified that the updated calcs. give values that are reasonable,
are not the same as the values given by the global stats AM, and are
not identical for both regions.

The same test case (on Blueskies) that was previously used can be used
for testing here.