Port changes made on ACME side to CIME for anlworkstations
This exposed some important errors. First, pio.py was not providing
the mpilib in the attribute list when ask for PIO data. This meant that
the selection of mpiserial was causing us to use netcdf instead of
pnetcdf as PIO_TYPENAME.

The second error was in entry_id. The algorithm for computing best match
was not picking up defaults in the case where the XML <value> tag has
no attributes.

Another error is that attributes were not being checked when loading
modules or setting env variables.