Tweak OS-specific setting in `config_compilers`.
In an upcoming refactoring, we will distinguish between information that
we have at setup time (e.g. the machine and OS) and information that we
have at build time (e.g. whether debug flags should be enabled). The
former should be on the top level `compiler` tags, so this commit moves
an OS tag accordingly.

This is not yet tested (ACME-specific change).