Further shrink `macros.py` output.
The file can be reduced considerably in size by leaving out settings for
compilers and MPI implementations that are not supported on the machine.

To implement this change, the `is_valid_compiler` and `is_valid_MPIlib`
functions had to be changed to occasionally return `False`, rather than
throwing an error in this case. However, it turns out that the error
being thrown was superfluous anyway, since the output of these two
functions was already being checked with an `expect` wherever they were
being called.