Move some CPP definitions from Python to Makefile.
`compilers.py` was setting `-D$OS` and `-D$COMPILER`, but the former was
redundant with a setting in the Makefile, and the latter is a bit
inelegant to do in `macros.py`, compared to simply having the Makefile
be responsible for this setting.

Compilers whose name does not match their vendor (e.g. `pgicuda`) still
need to define the compiler vendor in `CPPDEFS` in `config_compilers`/
`config_build`.