Process physics suites before setting up variable packages
The packages that are attached to physics variables in the Registry.xml
file depend on physics namelist options having their final name, after
the 'suite' name is substituted for the name of an actual physics option
(e.g., 'mp_wsm6'). Accordingly, we need to ensure that physics suites are
processed before variable packages are set up.

This commit moves the call to physics_namelist_check(), which processes
physics suites, from atm_mpas_init_block() to atm_setup_packages(). Also,
the 'mesh' argument to physics_namelist_check() has been removed, since
it was not needed and is not available in atm_setup_packages().