Optimize physics_addtend
As was done with other dynamics routines, apply the trick of dereferencing
array pointers through a subroutine call to a new routine physics_addtend_work( ),
and use constant inner array dimensions.

This commit also changes pointers to non-stride-1 sub-arrays of scalars and scalars_tend
to pointers to the entire var_array with use of "index_???" indices.