Adjust rx1 over 20 cells in QU_120km/with_land_ice
This merge increases the range over which rx1 is allowed to change
in the QU_120km/with_land_ice test case from 10 cells to 20 cells.
With the new init method, 10 cells is not enough to match up cleanly
to the z-level grid.

Also, adds verticalStretch as an output variable in init mode for
all global_ocean cases (via the template).