Changing MCT algorithm used to count number of active pes
The complexity of the existing algorithm in active_pes_ is
\Omega(ngseg * numprocs) in many instances, and can perform poorly at
scale (large ngseg and large numprocs). The new algorithm is
O(ngseg + numprocs).

This has been tested with a GMPAS-NYF compset with a resolution
of oRRS15to5 using 16384, 32768, and 65536 MPI processes on Titan.
There is however one option in active_pes_ that may not be exercised
in ACME - returning a pe_list. This capability has been implemented
(original method is not compatible with new 'count' algorithm), and it
compiles, but it has not yet been tested. A PR for this change has
been submitted to the MCT repository, and full testing should take
place for that PR.