Fix logic in warning about untested vertical level distribution parameters
When setting up als, alt, and zetal, we want to print a warning message when
the number of vertical layers is not either 41 or 55, which have been tested.
The logic controlling the printing of the warning message was previously
activated when (nVertLevels /= 55 .or. nVertLevels /= 41); however, this is
always true, causing the warning to be printed even when the user selected
either 41 or 55 layers. This commit simply changes the disjunction to
a conjunction, and cleans up the formatting of the code section slightly.