Fix vertical bounds check in rx1 constrained init
Fix a bug where rx1 constraint was not being applied properly to
the top 3 layers because maxLevelCell was set to 1 (as a placeholder).
Now, nVertLevels+1 is used as the placeholder and switched to
nVertLevels if the bottomDepth is not found.

Also, add computation of the max rx1 in each level (for debugging).