Adjust 'none' velocity solver to be based on uReconstructX/Y
Previously, the 'none' (data) velocity solver assumed normalVelocity was
input as data.  However, I realized that is awkward to set up as data,
and it is much more useful to pass in uReconstructX/Y and then have the
velocity code calculate normalVelocity.  Note normalVelocity for 'none'
velocity solver is really only needed if intending to do advection with
a data velocity field.