Changing PIO1 flow control logic for io2comp and comp2io
Porting Pat Worley's changes (from his branch) to CIME. This
commit changes the defaults of some PIO1 flow control runtime
parameters.

For io2comp, the handshaking protocol is very expensive (and
otherwise unneeded) for large process counts. The handshaking
protocol is however very important for comp2io.

This modification disables handshaking for io2comp when
FLOW_CONTROL is specified, to prevent this bad choice
from being enforced.

Specifying no limit on the number of preposted receive requests for
comp2io is also potentially a problem. (It is less of an issue for
io2comp, so is not addressed here.) While this can be avoided in
env_run.xml, this modification also adds logic to change the
'no limit' setting for comp2io to

      pio_maxreq = max(2*IOsystem%num_iotasks,DEF_P2P_MAXREQ)

as a small multiple of IOsystem%num_iotasks has proven to perform
well empirically. We can improve this further by communicating between
iotasks first, and then non-iotasks to iotasks, but that is a future
task).

[NML]
[BFB]