Remove `os_` from `MacroMaker` constructor.
The constructor already requires a `Machines` object, instantiated for a
particular machine, to be added. The operating system can be looked up
using this object, so there's no need to store it as a separate string.
(If the lookup ever becomes a performance bottleneck, we can always
cache the string.)