Fix sign on diffusive contribution to divergence
I was able to definitely test the sign on the advective and diffusive
contributions to divergence.  I set up a dome test case with no initial
waterThickness and then add meltInput only at the center cell.  I then
ran it with either the advective or diffusive contribution to divergence
commented out.  Success was when the divergence in the center cell was
positive (draining) and the surrounding cells were negative (filling)
with 1/6 the magnitude of the center cell.  Through this process, I
learned that the advective contribution has the correct sign, but the
diffusive contribution had the wrong sign.  This fixes that.

Note I also learned that edgeSignOnCell is positive to the NE for
periodic_hex meshes.  Well, more specifically that is how I am using it.