Translate `configure` script to Python.
At the time of this commit, the script itself has been tested, but the
changes to other Python modules have not. Though the output of the
script is somewhat different, the various tools in `tools/` have not yet
been changed to match the new format.

Note that the `configure` script uses `config_build.xml`, not
`config_compilers.xml`.