Add waterFluxDiffu variable
This adds separate variables for the water flux from advective and
diffusive components, plus the sum of them as the old waterFlux
variable.  It also moves the calculation of the diffusive flux to the
same place as the advective flux.  These changes make the two
calculations easier to follow and allows easier debugging because they
both can be output.

There are some roundoff level changes, but answers are not changed
meaningfully.