Updates CVMix tag in MPAS
This PR updates the version of CVMix utilized in MPAS-O.  A glaring bug
was discovered in CVMix that resulted in artificially enhanced diffusivity
in the layer directly below the boundary layer.  The new version of
CVMix fixes this issue.  When this issue has been fixed, artificial noise in
boundary layer depths is found. Utilizing enhanced diffusivity from Large
et al. 1994 (implemented in this new CVMix tag) removes this noise.  A flag
to enable this artificial diffusion has been added to the registry.  It should be
enabled in nearly all runs and is thus the flag is set to true by default