More changes to hydro-ramp - results close to desired
Adjusted the time step, run duration.  Note that using longer time steps
(which are allowed by the CFLs) leads to strange (though not necessarily
wrong?) behavior, so I've capped the adaptive dt at 1 month.

I've also noticed that the porosity value strongly influences the final
pressure field.  It's not clear to me yet if this is a bug or if the
pressure field takes longer to reach SS when porosity is large.  (It
does not appear to be changing much...)  For now, I've found that using
0.001 leads to close to deisred profiles, but further investigation is
warranted.