Prevent salt fluxes from removing too much salt
This commit modifies the computation of salt fluxes within the
computation of frazil fluxes and the bulk forcing module. Salt is only
removed if it is present, rather than removing salt regardless of how
much is present.

Surface fluxes are scaled back if the salt flux requires too much salt.

Additionally, arrays are added to keep track of how much salt flux is
ignored, and the actually salt content of frazil as it's formed.