Zero diffusive flux at margins - may be wrong
Calculating d(waterThickness)/dx at edges on the grounded ice margin
leads to seemingly unrealistically large gradients because
waterThickness is prescribed to be 0 where there is not grounded ice.

This commit zeroes that derivative, and therefore the diffusive flux at
those locations to avoid those unrealistic values.  However, I'm not
sure if this is correct either, so this may need to be revisited at a
later time.

Note that so far, the diffusive flux is << than the advective flux in
all the tests I have done.  This may be a bug or it may be reality - I'm
not sure yet.