Break meltInput into basalMeltInput and externalWaterInput
Depending on how basal melting is handled, it may be necessary to
distinguish between melt occurring at the bed and water entering the
subglacial drainage system externally.  This commit breaks the existing
meltInput field into one for the basal melt and one for the external
input.

Note that eventually we will basalMeltInput to come from the thermal
code (or optionally come from it).

With these changes, the hydro-ramp test case now matches Hewitt 2011
Figure 2 - the first completely successful verification of the model!