Improve reporting of mismatched dimensions in mpas_atm_dimensions module
The mpas_atm_dimensions module is responsible for either setting the dimensions
nVertLevels, maxEdges, maxEdges2, and num_scalars based on compile-time constants
or on input data at runtime. If compile-time constants are specified and do not
match the runtime values, the module previously printed messages to unit 0
and then called mpas_dmpar_global_abort. Often, an MPI task other than task 0
caused the job to abort, and a user would not be able to see the error message
about the mismatched dimension(s) in the log.0000.err file.

Now, the mpas_atm_dimensions module reports the compile-time and runtime values
for all dimensions it handles to the log.????.err files as well as to a brief
message that shows up in the critical error files created by mpas_dmpar_global_abort.