Change usage of deltat throughout model
Previously there was a local variable starting in mpas_li_core called
dtSeconds that controlled the model timestep.  Then there was an output
variable called deltat.

This commit eliminated dtSeconds and replaces its usage everywhere with
deltat.  This is to simplify things so there is only one time step size
variable.  Note there is a slight complication in that now this variable
has to be retrieved from a block, but all blocks always have the same
value.