Correct SGH adaptive time stepper to correctly relate to model time
Up until now, the SGH adaptive time stepper would not actually change
the model time.  It worked correctly, but the time in, e.g, xtime, was
not correct for how long elapsed in the SGH model.  This commit fixes
that by forcing the SGH adaptive time stepper to hijack the primary
model time variable, deltat.  In the future we probably want the SGH
model time stepper to subcycle underneath the master model deltat, but
this is fine for now.

As part of this change, a new config has been introduced:
config_SGH_max_adaptive_timestep
Previously this functionality was coming from the master model deltat
setting, but that no longer will work if the SGH model is changing that
variable.