hydro-radial test case - update values to match Bueler, big change to k!
This changes the setup script and some parameter values to better match
Ed Bueler's test case.  There is a very significant change to k
(config_SGH_conduc_coeff).  In Ed's Matlab code, k is K with units of
m/s so k in MPAS is k=K/(rhow*g).

With this change, MPAS reasonably matches the steady state solution!
Note however, this is for sliding all the way to the origin - later work
will need to extend this to the no sliding in the center configuration
used in the Bueler and van Pelt paper.

Also, there still is an issue with how to deal with water flux, and
subsequently thickness, around the boundaries.  If the hydropotential
gradient is artificially small there because of discretization error,
the flux is small or 0, leading to small or negative divergence, and
water piles up there.  I will look into this issue soon.

Followup commits will include other pieces of this test case needed to
reproduce the test case in the Bueler and van Pelt paper.