hydro-radial: fully working
This makes additional changes to hydro-radial test case and gets it to a
state where it can reproduce results from the Bueler and van Pelt paper.

* the correct sliding profile is turned on.
* the run duration and porosity values are set to the correct values.
* a README is added that fully describes the test case and how to run it
  (many steps currently have to be run manually)
* the plotting script is updated

Note that while MPAS is able to match the Figure in the GMD paper, there
still are some kind of strange results from the model.  It may be that
the short 1-month test is not strict enough to identify them, or it may
be that they aren't a big deal.  Specifically, waterThickness in
marginal cells increases unrealistically because the hydropotential
gradient out of the last cell is too weak due to discretization error.