Add halo update to divergence - decomp test now passes!
With this change, the hydro-radial decomposition test now passes!

I tested removing the halo update on waterFlux after adding the halo
update on divergence but that did not work.

I also tested the number of halos required for this configuration
(halo updates on both waterFlux and divergence) using the decomp test,
and found only 1 is needed.