Improve logic about velocity solver-specific calculations
A bunch of calculations were being done unnecessarily when the 'none'
velocity solver was selected.  This changes some logic in the
diagnostic_variable_before_velocity routine to check for higher order
velocity solvers explicitly, rather than just checking for not SIA.

Also, added a timer to the mask calculation that reveals it is
expensive!