Switch from 'index(name, "VAL") /= 0' to 'trim(name) == "VAL"' in init_atmosphere core
The old method of comparing names of fields in intermediate files to determine which block
of code is used to handle the interpolation of the field led to false matches, e.g., "HGTTROP"
matching "TT", or "SNOWH" matching "SNOW". So, we switch to a more robust method of comparing.

Note: This commit does not replace uses of 'index(...)' in mpas_init_atm_hinterp.F or
mpas_init_atm_llxy.F, but only in code that matches against the names of field in intermediate
files.