Improve mask calculation performance
A number of mask items are only needed by the Albany velo solver.
Checking if Albany is active within the vertex loop in the mask
calculation made it more than 10x more expensive than the cell and edge
mask calculations, even when Albany is inactive!

This commit moves those optional Albany-related vertex mask calculations
into a separate vertex loop.  This means the check if Albany is active
only needs to be evaluated once.  This reduces the vertex mask
calculation cost by an order of magnitude, making it comparable in cost
to the cell and edge mask calculations.

I also added commented out timers for the cell, vertex, and edge mask
calculations should further optimization be desired.