Dome restart test - clobber_mode=truncate for cold start
Previously, the clobber_mode was overwrite for this restart test.  But
this meant that you had to manually delete the output file to ensure
that you were using a clean file when redoing the test.  By changing the
coldstart part of the restart run to truncate, it ensures that the
output.nc file is new to that run and not some old junk lying around.