Avoid temporary array creation with ifort for 'qvapor' in compute_slp()
The 'qvapor' field was previously set as

   qvapor  => scalars(index_qv,:,:)

and passed to the compute_slp() routine, which uses qvapor twice. However,
the ifort compiler issues warnings like

   forrtl: warning (406): fort: (1): In call to COMPUTE_SLP, an array temporary was created for argument #6

   Image              PC                Routine            Line        Source
   libmpi.so.12       00002B4FC30A7B73  Unknown               Unknown  Unknown
   atmosphere_model   0000000000A2F13F  mpas_atm_interp_d         545  mpas_atm_interp_diagnostics.F

when the code is compiled with DEBUG=true. To avoid the creation of
a temporary array for 'qv', we simply pass the entire 'scalars' array,
along with 'index_qv', into the compute_slp() routine.