Set RK substeps and acoustic steps consistenly for 2nd-order time integration
The computation of the coefficients for the vertically implicit solve were
moved above the RK loop during optimization, and each acoustic step should
therefore be the same length. For ns=2, the acoustic steps were properly set
to dt_dynamics/2, but the first RK step size was still set to dt_dynamics/3.

This commit sets the first RK step to dt_dynamics/2 to give 2nd-order integration.