Create SHMIP test case - just test A for now
Subglacial Hydrology Model Inter-comparison Project (SHMIP)
These tests are described here: http://shmip.bitbucket.org/

This commit begins to add these tests.  There are organized in a way
where as many of the tests as possible get set up, but all the tests
must be run manually.

This commit adds test A, including a script to plot the results of test
A3 against a benchmark from the GLADS model provided by the
intercomparison.