Avoid divide by zero in effectiveConducEdge when hydropoten. grad. is tiny
When beta<2, small gradients in the hydropotential result in very large
effectiveConducEdge and, in turn, diffusivity, which leads to very small
time steps.

This commit adds a lower limit on the hydropotential gradient, below
which effectiveConducEdge is just set to 0.  However, this is somewhat
arbitrary and may not be appropriate in all cases.  Also, diffusivity is
still quite large where the hydropotential gradient is above this
threshold.