Remove duplicated call to mpas_atm_soundings_finalize() in atm_core_finalize()
Somehow the call to mpas_atm_soundings() was duplicated -- perhaps in the merge
of the 2016 branch onto the 2015 branch. This duplicated call caused us to try
to deallocate the same arrays twice.

Ideally, this merge should be squashed into the merge that introduced the duplicate
call, and a fix in the mpas_atm_soundings module should be made to ensure that
it is not possible to double-deallocate memory.