Remove nested k loops in cvmix interface for kpp
This commit removes three nested k loops within the cvmix/kpp interface.

Two of these are using a sum() intrinsic within k loops (over some
subset of the k levels). The last of these is removing the bulk of a k
loop that computes the surface layer for averaging over.