Add landIcePressure variable
Land-ice pressure is now added to the pressure at the surface, but
is separate from the seaSurfacePressure.

The iterative initialization now works with ssh and landIcePressure
(rather than seaSurfacePressure) and has been renamed accordingly
(both the module and the namelist).

A new mode for land ice fluxes has been added, 'pressure-only',
which allows the model to run with landIcePressure but without
the other fluxes related to land ice.  This is useful for iterative
initialization of ssh and landIcePressure, for example.

Many test cases have been updated both in init mode and in the
test_cases infrastructure to work with landIcePressure and the
associated namelist changes.  The scripts for iterative initialization
and plotting have also been updated to work with landIcePressure.