Change mpas_io_streams to use dminfo from ioContext instead of block
This commit changes mpas_io_streams to use the dminfo object from a
stream's ioContext, rather than from a fields block / domain. This is
useful when writing debugging streams out in a way that uses a different
dminfo.

A helper function is added to mpas_io to allow the mpas_io_streams
module to access the dminfo object that a fileHandle uses, without
directly accessing it.