Change implementation of default caseroot for check_lockedfiles
For some reason, some tests in scripts_regression_tests were failing for me with
the old version of this code - maybe because default arguments are evaluated
when the function is defined (not when it is called)?

This new version seems more likely to be what's wanted - i.e., getting the
current working directory at the time when the function is called, not the time
when it's defined.

In particular, this seems like it could (in principle) be important in
case_setup, which does:

    os.chdir(caseroot)

before calling:

            check_lockedfiles()

Test suite: None
Test baseline: N/A
Test namelist changes: N/A
Test status: N/A

Fixes: None

User interface changes?: No

Code review: None