Adding area correction based on 'epsilon bubble' for RRM meshes
*********1*********2*********3*********4*********5*********6*********7**
This is not related to area corrections for metric terms but:
I noticed some unused variables and compiled with ifort on skybridge with flag '-warn unused'. There are screens of warnings of unused variables in pio, timing, and homme. This is an attempt to clean some of them from files listed below.

Variables in #ifdefs and variables in some disabled code for debugging or reference were not cleared. In more obvious cases where the variable is really never used the variable is deleted. There is no difference in homme suite output for this commit (cleaned vars) compared to the previous commit (modified metric terms). Also, one of SMS tests ran (but wasn't compared to baseline). There are many more cases of unused vars to go through. I will create a branch.

	modified:   components/homme/src/share/coordinate_systems_mod.F90
	modified:   components/homme/src/share/derivative_mod_base.F90
	modified:   components/homme/src/share/edge_mod_base.F90
	modified:   components/homme/src/share/gridgraph_mod.F90
	modified:   components/homme/src/share/ll_mod.F90
	modified:   components/homme/src/share/metagraph_mod.F90
	modified:   components/homme/src/share/viscosity_mod_base.F90