Arbitrary vertical interpolation of global ocean initial condition
Previously, the input data initial conditions (IC) had to be on the
same vertical grid as the final grid produced in init_step2.  This
PR includes the following:
- independant vertical grids for activeTracer IC, ecosystem IC, and
  final IC.
- new test case, QU_240km/zstar_128_layers, makes use of interpolation
  from 100 layer IC data files to 128 layer MPAS IC file.
- substantial clean-up of mpas_ocn_init_global_ocean.F
- unification of interpolation and smoothing routine for active tracers
  and ecosystem tracers (was two routines before)
- read in data IC for each tracer separately.  Before, T&S were within a
  loop, so we would read in data T&S IC together and sometimes run out
  of memory.

This PR prepares for sigma and hybrid coordinate initialization, but
does not yet implement them.

This PR is bit-for-bit against ocean/develop using the nightly
regresssion test, except for 'with land ice', which was close.  The
QU_120km/bgc_60_layer was also tested and works.