Bug fixes for dynamic root model
This commit message updates the dynamic root model to fix the
spinup delay in the tropics. The updates include a revised
water profile that matches btran, a exponential N profile when
vertical profiles are not on, and allows changes to the root profile
during the growing season rather than just the onset period for
stress and seasonal deciduous vegetation.

[Non-BFB] - Non Bit-For-Bit