Add dissipation terms to cpp interface, not yet passed
These changes create an array on the C++ side of the interface that
provides the heat dissipation as calculated by Albany on MPAS vertices.
This code was written by Mauro Perego.

Currently this field is only created on the C++ side of the interface -
it will be passed to MPAS in a follow up commit.