Pass Albany dissipationHeat through interface
This modifies both sides of the interface to pass dissipationHeat back
to MPAS.  Nothing is yet done with it on the MPAS side.  It is passed
into a scratch variable and will then be interpolated onto cell centers.
Until that interpolation occurs, I have left the scratch variable
'workLevelVertex' as 'persistent' so it can be visualized.