Fix crit. pass. and file_interval, RRS_30to10km/spin_up
Link to critical passages geometry has been removed because none
was created in init_step1.  Critical passages are now not deepened.

In the RRS_30to10km/spin_up test case, the restart file in the
spin_up2 step isn't read correctly unless the file interval is
reduced to the 4-hour interval of the spin_up1 step.