Working around implicit array segment copies
In the routine advance_clubb_core (in
cam/src/physics/clubb/advance_clubb_core_module.F90) there are loops
of the form:

do k = 1, gr%nz, 1
 call pdf_closure &
  (...
   zm2zt( wpthlp, k ), rtpthlp_zt(k), sclrm(k,:),    & ! intent(in)
   wpsclrp_zt(k,:), sclrp2_zt(k,:), sclrprtp_zt(k,:),& ! intent(in)
   sclrpthlp_zt(k,:), k,                             & ! intent(in)
   wphydrometp_zt(k,:), wp2hmp(k,:),                 & ! intent(in)
   ...
   rtphmp_zt(k,:), thlphmp_zt(k,:),                  & ! intent(in)
   wpsclrprtp(k,:), wpsclrp2(k,:), sclrpthvp_zt(k,:),& ! intent(out)
   wpsclrpthlp(k,:), sclrprcp_zt(k,:), wp2sclrp(k,:),& ! intent(out)
   ...
  )

Each of the 15 arrays of the form XXX(k,:) is declared internally as
an array of size XXX(:), and the compilers apparently are creating
local temporaries and copying into and out of these. This is pretty
low level (being inside loops over first chunks, then local columns,
and then nadv).

Explicitly allocating temporary arrays of the correct dimensions and
copying into (for intent(in)) and out of (for intent(out)) external to
the call to pdf_closure improves performance.

For the Intel compiler on Titan, this drops the cost by around 15%.
For the PGI compiler on Titan, this decreases the cost by a factor of
6.

This modification only modifies two of the loops containing calls to
pdf_closure, as these are the only two that are exercized in the
current ACME test cases. There are two others that should be modified in
analogous ways if l_use_ice_latent is true.