Change reference_time 0001-01-01
Because start_time defaults to 0001-01-01, the reference time for
all registries and test cases has been changed to be the same. This
prevents a problem where output and restart files are getting
unexpected date stamps.  For example, if start_time is 0001-01-01
but reference_time for restarts is 0000-01-01 and the output_interval
is 10 days, the first restart file was getting stored with date
0001-01-06 (370 days after the reference date) but with the date
stamp 0001-01-11 in xtime and in the Restart_pointer.

Reference times have been added to templates for several analysis
members that didn't explicitly include them but where the file
names include the date and time (thus, the same confusion could arise).